+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Update `ActivityStoreContext.isFragmentStarted` and `ActivityStoreContext.isFragmentResumed` functions.
 - **Breaking**: Removed `BackStackStore` and `BackStack`.
 - **Breaking**: Removed `SharedStateStore`.
+- **Breaking**: Replacing `FlowDeclaration` and `FlowIntegration` with `FlowFactory`.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.